Rename persistence/http to persistence/api
